continue {circumstance}	R Documentation

Continue an iterative calculation

Description

Continue an iterative computation where it left off.

Usage

## S4 method for signature 'mif2List'
continue(object, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)

## S4 method for signature 'pmcmcList'
continue(object, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)

Arguments

object
the result of an iterative pomp computation

...
additional arguments will be passed to the underlying method. This allows one to modify parameters used in the original computations.

seed
seed for the parallel random-number generator. Setting seed=TRUE passes the current seed to the parallel RNG. See %dofuture% for more information.

chunk.size
average number of elements processed per chunk. See %dofuture% for more information.

scheduling
average number of chunks that each worker processes. See %dofuture% for more information. This is ignored unless chunk.size=NULL.

See Also

mif2

[Package circumstance version 0.0.10.1 Index]